Renamed unit test files for actions
